playto-pay/backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.12
celery==5.3.6
redis==5.0.3
django-cors-headers==4.3.1
pytest==8.1.1
pytest-django==4.8.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1

